requirements.txt
python-can>=4.3.1
cantools>=39.2.0
pyserial>=3.5
PyQt5>=5.15.10
pyinstaller>=6.15.0
cryptography>=41.0.7
numpy>=1.24.3
pandas>=2.0.3
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0 